requirements.txt.txt
discord.py  
python-dotenv  
apscheduler  
pytz